refactor: move downloader out of sync module
